grassdata conversion scripts added
This PR adds a set of GRASS GIS 8 scripts used to convert the large GRASS GIS 6.4 database that CASAS has built so far to GRASS GIS 8.
It converts raster data as well as vector data (topology update as well as attribute transfer from DBF to SQLite).

A `README.md` is included, describing the workflow.